fixing warnings
git-svn-id: svn://localhost/neomemC/trunk@353 3bad96a7-2d8f-4f48-a7ae-748201937164

+++ code/BDataIcon.cpp
@@ -201,7 +201,8 @@ BOOL BDataIcon::LoadFile(LPCTSTR szFileName)
 	BOOL bOK = f.Open(szFileName, CFile::modeRead, &fe);
 	if (bOK)
 	{
-		m_nBytes = f.GetLength();
+		// cast is okay, assuming icon file < 4GB
+		m_nBytes = (UINT) f.GetLength();
 
 		// Allocate chunk of memory and read entire file into it
 		m_pData = new unsigned char[m_nBytes];
